// types/next-auth.d.ts
import NextAuth, { DefaultSession } from "next-auth";

export interface Credentials {
  username: string;
  password: string;
}

declare module "next-auth" {
  interface Session {
    user: {
      id: string;
      username: string;
      role: string;
    } & DefaultSession["user"];
  }

  interface User {
    id: string;
    username: string;
    role: string;
  }
}

declare module "next-auth/jwt" {
  interface JWT {
    id: string;
    username: string;
    role: string;
  }
}